requirements.txt
pandas>=1.3.0
spacy>=3.0.0
scikit-learn>=0.24.0
numpy>=1.19.0 
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.1